# openapi.yaml
openapi: 3.0.0
info:
  title: 꼬리콥터 API 명세
  description: 꼬리콥터 백엔드 API 명세서입니다. 초기 온보딩, 산책 기능, 코스 및 기록 관리를 위한 API들을 포함합니다.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: 개발 서버
...
paths:
...
  /courses/recommendations:
    get:
      tags:
        - 산책 기능
      summary: 우리 동네 추천 코스 요약 목록 조회
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
        - name: longitude
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          required: true
          schema:
            type: integer
        - name: sortBy
          in: query
          required: true
          schema:
            type: string
        - name: areaName
          in: query
          schema:
            type: string
        - name: petSize
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 추천 코스 목록 반환
  /courses/{courseId}:
    get:
      tags:
        - 산책 기능
      summary: 선택된 추천 코스의 상세 정보 조회
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 코스 상세 정보 반환
...
  /courses/new:
    post:
      tags:
        - 코스 및 기록 관리
      summary: 새로운 코스 정보 최종 등록
      requestBody:
        description: 새로운 코스 정보
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walkRecordId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                imageUrl:
                  type: string
                recommended:
                  type: boolean
      responses:
        '201':
          description: 코스 등록 성공
  /courses/new/details:
    get:
      tags:
        - 코스 및 기록 관리
      summary: 새로 생성할 코스의 기본 정보 조회
      parameters:
        - name: walkRecordId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 코스 기본 정보 반환
  /courses/{id}/marking-photozones:
    get:
      tags:
        - 코스 및 기록 관리
      summary: 코스 포토존 목록 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 포토존 목록 반환
...